requirements_backend.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
supabase==2.3.0
google-genai==0.1.0
